rename tourou molecule
